+++ src/components/animate-scramble/index.ts
@@ -1,0 +1,2 @@
+export { MPAnimateScramble } from './MPAnimateScramble';
+export type { MPAnimateScrambleProps } from './MPAnimateScramble';